Extract trace writer masking helpers
